Fix fullAddress, Identifier, hasPrizeRank
Fix multiple fullAddress values

- Apply the fullAddress technique to the ContractTerm definesSpecificPlaceOfPerformance, which was missed. [fixes #19]

Remove unnecessary ContactPoint identifier

- OPT-201-Organization-TouchPoint need not be mapped as ePO does not
foresee an Identifier for the ContactPoint. [fixes #15]

Fix hasPrizeRank multiple values, unmap BT-45-Lot

- The template path for the Prize subject was incorrect (at the wrong
level), and therefore its predicates had repeated values. [fixes #16]

- As an indirectly related issue that has not yet been reported or
discovered: BT-45-Lot creates a discrepancy as descriptions are in the
prizes, which ePO does not foresee, and so the current mapping to the
DesignContestRegimeTerm's hasFollowupContractInformation is not
correct.