Add versioned packages EF10-24 SDK v1.3-1.8+1.0
RML rules are versioned into version-specific RML files/modules with a
version range, of the form:

- Concept (common/shared)
- Concept_v1.x-1.x (specific version x)
- Concept_v1.x-1.y (version range x to y, inclusive)
- Concept_v1.x+ (minimum version x, inclusive)
- Concept_v1.0-1.x (no minimum version)

Fields with backward-compatible, non-functional, stylistic or, in
general "superficial", changes in the XPath are documented in code on a
best-effort basis.

For non-backward-compatible changes, separate versioned files are
created following the above file naming scheme. The best effort is made
to keep as much shared code as possible within the common file, such as
the relevant TriplesMap with the subject template. However, at a
minimum, the TriplesMap label is kept, in order to provide a versioned
indication. In other cases, the TriplesMap is moved wholesale to
versioned files.

Among RML rules in versioned files, renames of the node name component
of a TriplesMap is avoided so as to maintain the most compatiblity with
referencing TriplesMap (via `parentTriplesMap`), where possible. For the
same reason, iterator changes just to match an SDK node change are
avoided, as the change in XPath is often superficial (effective XPaths
are the same).