Fix gh-31 undesired repeated role instantiation
Fixes instances of the following 10 epo:AgentInRole subtypes to not
repeat for the connected Lots:

- Reviewer
- EmploymentInformationProvider + fix duplicate and wrong mappings [1]
- EnvironmentalProtectionInformationProvider + remove duplicate mappings
- Mediator
- OfflineAccessProvider
- ProcurementProcedureInformationProvider
- ReviewProcedureInformationProvider
- TaxInformationProvider
- TenderProcessor
- TenderReceiver

[1] Not only was there a duplicate `announcesRole` in the same
TriplesMap, but there was an entirely duplicate TriplesMap for
`contextualisedBy` somewhere it shouldn't belong, and with an
incorrectly linked parent with the wrong iterator. This wouldn't have
been caught as results look almost the same as in this fix, except with
a hidden bug (missing basic link to the Lot ending up creating a
connection to all Lots).