Fix gh-52 missing hasLegalBasis for certain non-EF16 CN notices
The predicate mapping was in the wrong TriplesMap (TMap) that has an
incompatible iterator (a toplevel `TenderingProcess`, which is not
guaranteed to exist), and an absolute XPath reference. It should've been
in the root TMap which has the right iterator (simply the root of the
XML document `/*`, which also allows a relative XPath in the mapping).

Affects EF14, EF15 and EF19 as reported. Tested with EF14 notice
152054-2024.xml.

The following cases have data (from Official Journal publications)
included for testing:

- EF14 for SDK version 1.8 (00152054-2024)
- EF15 for SDK version 1.7 (00140913-2024)
- EF19 for SDK version 1.10 (00128154-2024)

As downloaded from URLs of the form
https://ted.europa.eu/en/notice/152054-2024/xml (first example).

Closes gh-52.