Add additional example data in CNs - invalids and 100_lots
Invalids were not included since D2.2 (versioned CNs),
cn_24_maximal_100_lots.xml was never included but is still part of the
SDK for CNs, so should be included in the packages for consistency.

Note that *100_lots.xml is a large file and may take significantly
longer to transform depending on compute resources.