Add output and validation assets for CN MS v1.9 D2.2 parity
As part of the update to CN v1.9, bringing D2.1 up to parity with D2.2
versioned CNs (v1.3-1.8+1.10).

Assertion changes are merely renames as they are the same content with
different file names due to newly generated IDs, so can be ignored.

The outputs can be observed to be reflecting the changes that went into
the last D2.2 redelivery, showing several fixes to instantations, e.g.
the AgentInRole ones for OPT-301-*, and the "collapsing" of other
erroneously duplicated instances.