Add output and validation assets for CN MS D4.1 parity
As part of the update to all CN versions, bringing D2.2 (rc.6) up to
parity with D4.1 CANs (v1.3-1.10) involving CN rules.

Assertion changes are merely renames as they are the same content with
different file names due to newly generated IDs, so can be ignored.

The outputs can be observed to be reflecting the changes that went into
CN rules of CAN packages in the last D4.1 delivery (rc.7), showing some
fixes to instantations, e.g. MonetaryValue for Procedure instead of Lot,
and the "collapsing" of erroneously duplicated instances of
StrategicProcurement.