Add invalid example data in CANs
Invalids were not included since D2.2 (versioned CNs), but were
reintroduced in the last revision for CNs. They add some unwanted noise
to reports, but we're adding them for consistency's sake.

Tooling should be adjusted to take these into consideration in a special
way (negative assertions), so that reports become meaningful, the same
way valids are.