Port runts test as replacement for empty instances validations
There were two separate validations for empty entities -- one in
genericEPO and the other in genericRDF (was: genericSPARQL). They did a
similar thing, except in the latter case the test was on ePO classes
specifically. This is unnecessary and we can have a truly generic one,
which is simpler (does not do any counts).

P.S: A "runt" is a special case of a "widow" -- a referenced entity that
has zero information (a runt has at least some albeit useless data, like
an `rdf:type` assertion).