Add BT-766-Part versioned mappings
The node changed in SDK v1.9 but we continue using the newer iterator
(newer is more restrictive), while moving only the predicate mapping to
relevant versioned files, as the relative XPath is different (the older
one is more restrictive).

In effect, this is a redundant versioned mapping implementation, meaning
it is not pragmatically required as the XPaths in the different versions
play out nicely in terms of the later versions being more restrictive,
but we nevertheless choose to do it in this case for demonstration
purposes.