Fix specific/broad place instantiation, refactor contract terms
- Reintroduce check for country, fixing missing location in some cases
  (e.g. `pin-only_25.xml`)

- Guard against `cbc:Region` to not create an Address for a broad place
  of performance (the specific place will nevertheless be created but
only to declare the country)

- Fix the place of performance contract's subject template to
  instantiate at the `cbc:RealizedLocation`, creating a separate
contract and harmonizing all place-related information together

This fixes a long-standing issue of the Contract Term vs. related Place
of Performance instantiation, where the descriptions for both the broad
and specific places were asserted together.

WARNING: These changes and improvements apply only to PIN, but a similar
refactor needs to take place for the CN and CAN mappings of related
fields if the place additional information multiple values is desired to
be fixed. However, that would affect older mappings and therefore
transformation behaviour, so such an effort may best be deferred until a
later time.