Fix BT-1501(s)-Contract versioned mapping for v1.6 and below
Up to SDK v1.6 the `efbc:ChangedSectionIdentifier` is repeatable, and
therefore the key mappings for this field, which is on an ancestor
iterator, will fail. As the subject templates can contain conditions
with functions, XPath/XQuery processors cannot proceed due to receiving
more than one value for the function input, which is unexpected.

The fix is to iterate the TriplesMaps on the XPath that is repeatable,
essentially the XPath of the value itself. However, this results in
multiple values for `epo:relatesToEFormSectionIdentifier`.
Unfortunately, it is either that, or we reproduce the entire information
across two instances, which is not any more appealing.