Fix isProcurementDocumentChanged values, refactor ChangeInformation
The `epo:isProcurementDocumentChanged` had multiple values in some
cases, and this was a long-standing issue about instantiation relating
to the change XPaths differing across versions.

We employ a trick to instantiate at the right path in the template
depending on whether there is one or more Change elements. Furthermore,
we replicate the TMaps necessary for each unique node to facilitate the
different levels of instantiation.

This results in either one complete ChangeInformation, or several
instances of it with the description-related changes in one, and other
information in the other. This is because all of the information are not
structured in a sensible way in eForms, that would allow them to be
instantiated as one single resource in all cases.