Update language vocabulary resource files as of April 2025
The results are mostly reordering of items, as the queries do not
enforce an order, except the inclusion of OP_DATPRO.

Fix also UNIX file permissions for form-type vocab in src. The
permissions were already correct in the packages (their executable bit
is not set and not required as they are text files).