Zip up all reports to reduce project size, offer tooling compatibility
Some tooling like the notice processing pipeline, as of its current
incarnation, depend on the existence of the lowest level of reports.
However, they rely on only a partial matching criterion, like "sparql",
"shacl" and "xpath", and require the existence of 3 files of any
variant. It is therefore sufficient to have all file formats of each
variant of the report archived in a ZIP file.