Remove large example files for runtime limits compatibility
Tooling like the conversion pipeline, which writes to a MongoDB database
that has a per-document size limit of 16MB, will fail to load packages
that exceed that limit. Removing the `100_lots` example file allows us
to reclaim some space.

However, the root cause is actually the `language.json` vocabulary file,
which is relatively large. Although used previously successfully, it
just so happens that, when put together with these new combined
packages, the v1.10 package exceeds the limit.

We exclude the example file across all packages for consistency.
However, one may reinclude it manually for testing purposes or
otherwise.