Fix ExclusionGround type mapping for v1.12+, new vocab resource
The `dct:type` of an epo:ExclusionGround has used the `criterion`
controlled list (an EU Vocabulary "authority table", or AT) so far.
However, this is only applicable up to SDK v1.11.

From v1.12 onwards, a new AT `exclusion-ground` becomes applicable, that
is no longer a "sublist" of `criterion` and has the new `exg-` codes
used in newer notices.

Fixes TEDSWS-314.