Fix BT-67(b)-Procedure(-Language) with versioned mapping for v1.12+
There was a change in the XPath for BT-67(b)-Procedure that goes along
with the vocabulary change for BT-67(a)-Procedure. This triggered a
general RMLMapper error on the language mapping function, as it matched
the older XPath, which expects a language tag.

In such a case the notice may have failed to get through the
transformation, as not all tools will be aware that output is still
produced despite the error. Updating the XPath for SDK v1.12+ yields a
non-match for current notices, and therefore the error will not trigger.

Note that in the event there are any valid XPaths that are expected to
come with a language attribute, such an error will still be generated.
The underlying issue will therefore need to be solved in an appropriate
way (adding guards, reporting upstream, error-handling in tools).

Note: The error manifests in notices like 336207-2025 for this specific
case, and some SDK v1.3 example notices for the general missing-language
case. Also, the language TriplesMap has been kept in its original
formatting, though not ideal, as we may need to be able to trace it back
in its original form at a later time.

Fixes TEDSWS-359.