Fix BT-541-LotsGroup-ThresholdNumber rr:child -> rml:reference
There was a typo for an RML code whereby `rml:reference` was
inadvertently written `rr:child`, following the parentTriplesMap
instantiation terms even though this is a simple attribute mapping.

An RML processing error would be raised for any notice having the XPath
for this field, such as 173230-2025. This is now fixed and therefore
such notices will transform normally.