Guard against empty DynamicPurchaseSystemTechnique
In some notices the DPS is forbidden. Adding a guard solves this problem
for not only those notice subtypes, but all of them, where the
ContractingSystemTypeCode and its parent are absent. This can be seen in
the PMC reference example test file. In addition, we include for testing
a sample notice where this is forbidden, 2025-OJS002-00001899
(1899-2025).

Fixes TEDSWS-303.