Fix SelectionCriterion mapping for weight predicates in CN v1.3-1.8
The XPaths in the predicates related to weight were wrong according to
the iterator of the subject, for SDK v1.3-1.8. However, the versioned
mapping used a TMap shared with other predicates from a different node.
Therefore, we add a complete mapping in the versioned file with the
correct iterator.

Related to TEDSWS-320.