Fix empty ExclusionGround instances in PIN v1.12+
Move ExclusionGround subject TMap into versioned files, as the related
fields BT-67(a)-Procedure and BT-67(b)-Procedure have different XPaths.
There is now a listName attribute value `exclusion-ground-source` in
v1.12+ PIN notices, which is currently undefined in the SDK.

Catering for both cases in a base/shared TMap is problematic, as in new
notices the `@listName=exclusion-ground-source` is not matched, thereby
leading to runt ExclusionGround instances. Add appropriate conditions in
both versions to treat the situation consistently.

Additionally, remove an unwanted `rml:reference` from the BT-67(a) pOM,
and also adjust the label for these fields in v1.12+ according to the
SDK.