Fix SelectionTerm parameter mixup in SDK v1.3-1.8
The fields BT-752-Lot (weight) `cccev:weight` and co-occurent BT-752-Lot
(threshold) `epo:hasTresholdValue` were being confused for one or the
other in certain notices. Either the weight will be a duplicate of the
threshold, the threshold be mistaken for the weight, or the weight be
mistaken for the threshold. The effect can be hard to observe, because
numbers can often be the same for both, or they don't co-occur.

The fix lies in the conditional checks in the predicate mappings, where
the notation `if(exists(SOME_CONDITION/something='value')) then ...` is
not valid. The equality needs to be wrapped in an XPath filter/predicate
`[]` such that `if(exists(SOME_CONDITION[something='value'])) then ...`,
otherwise the condition would always return True (a value would always
exist, not being checked for what the value is).

Interestingly, in the recent fix for SelectionCriterion vs.
SelectionTerm for these properties (weight and threshold), some notices
would show a welcome change (introduction of `cccev:weight`), but in
others, for e.g. our SDK v1.8 example, the same change is observed
but is in fact unwelcome (in hindsight).