Fix Notice->LotGroup refersTo only for can, can-modif, compl
The mapping for Notice refersTo LotGroup was added as a blanket for all
notice subtypes in the root mapping file. However, according to ePO 4,
this relationship applies only in certain subtypes ResultNotice (in
root-can), ContractModificationNotice (in root-can-modif) and
CompletionNotice (in root-compl). Move the mapping accordingly and
update all sample outputs.