Fix missing AwardCriterion threshold Constraint linkage
This occurs due to an unwarranted connection to Constraint from
SelectionCriterion and missing AwardCriterion TriplesMap for
ND-LotAwardThresholdCriterionParameter `number-threshold`. Affects
BT-5423-Lot and BT-541-Lot-ThresholdNumber. Issue can be observed in
notice 775926-2023 (CN v1.10).

There is also a missing linkage for the
ND-LotAwardFixedCriterionParameter number-fixed counterpart, affecting
BT-5422-Lot and BT-541-Lot-FixedNumber, due to duplicate connection for
the ND-LotAwardWeightCriterionParameter `number-weight` counterpart.
This is not evident in the above notice but may be in others.

Fixes gh-154.