Fix wrong definesLotGroup iterator
The predicate mapping for `epo:definesLotGroup` was at a broad iterator
(at ND-LotDistribution cac:LotDistribution instead of
ND-GroupComposition cac:LotsGroup), where the mapping would fail if
there were to be more than one LotsGroups.

This is due to an upward traversal parent join which fails if there is
more than one element (in this case LotsGroup), in notices like
217262-2025. Create a new TriplesMap at the right iterator to mitigate
this.

Fixes gh-155.