Fix refersToPreviousNotice for notices without ChangeReason
The mapping for BT-758-notice has been mistakenly placed under the TMap
for ND-ChangeReason, which has a different iterator and context, mainly
to express the reason for a change. However, a ChangeInformation notice
like 207280-2025 may not necessarily have a ChangeReason, and therefore
the only right place where this should be declared is where
concernsNotice is -- at ND-Changes.

Fixes gh-156.