Fix specific/broad place instantiation for CN, CAN
This affects the eForms fields **BT-727-Lot** and **BT-728-Lot**, where
the instantiation of the ContractTerm and in turn the broad and specific
places of performance is not optimal, resulting in multiple values being
asserted for the properties
`epo:hasPlaceOfPerformanceAdditionalInformation` and
`epo:definesSpecificPlaceOfPerformance`.

This was accounted for during the mapping of the counterpart `-Part`
fields (for PIN notices), where the correct mappings were created to
avoid this issue. See commit fe7e3586be391bfe81ac898b12a2b8112db4ee4d
"Fix specific/broad place instantiation, refactor contract terms".
However, the issue still persisted in CAN and CN mappings.

A guard is placed against `cbc:Region` to avoid creating an address for
a broad place. The subject templates for the place of performance's
contract is updated to instantiate at the `cbc:RealizedLocation`. This
results in separate but harmonized contracts collecting place-related
information together. This in turn avoids multiple values being asserted
for the additional information and specific place.

Fixes gh-151.