feat: change URI scheme for external references and their identifiers
References to external entities are not limited to the notice. It can be
to a procedure, a review-related entity, or even something else
entirely.

In some cases, a reference is given without sufficient context. For e.g.
in the case of an external FrameworkAgreement only the external notice
ID may be given, but nothing about the contract itself. This therefore
violates our chosen scheme, because we cannot populate an external ID in
the trailing section of the URI:

```
{current-notice}/Notice/{external-notice}/{external-entity}/{external-id}
```

Even among notices, nested entities of the referred notice do not play
well with this URI scheme we have decided on. That is because, given
only two positions in the URI for entity/class names, it raises chances
for collisions for IDs in the trailer.